Move code to create tag files to files.rs
Moves the code to create tag files from main.rs to a new function
create_tag_files in files.rs.
